+++ packages/baseten/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/baseten
 
+## 1.0.0-beta.37
+
+### Patch Changes
+
+- Updated dependencies [db913bd]
+  - @ai-sdk/provider@3.0.0-beta.17
+  - @ai-sdk/openai-compatible@2.0.0-beta.35
+  - @ai-sdk/provider-utils@4.0.0-beta.34
+
 ## 1.0.0-beta.36
 
 ### Patch Changes