+++ packages/llamaindex/CHANGELOG.md
@@ -1,5 +1,84 @@
 # @ai-sdk/llamaindex
 
+## 2.0.0-beta.114
+
+### Patch Changes
+
+- Updated dependencies [69768c2]
+  - ai@6.0.0-beta.114
+
+## 2.0.0-beta.113
+
+### Patch Changes
+
+- Updated dependencies [79a8e7f]
+  - ai@6.0.0-beta.113
+
+## 2.0.0-beta.112
+
+### Patch Changes
+
+- Updated dependencies [e06b663]
+  - ai@6.0.0-beta.112
+
+## 2.0.0-beta.111
+
+### Patch Changes
+
+- ai@6.0.0-beta.111
+
+## 2.0.0-beta.110
+
+### Patch Changes
+
+- Updated dependencies [db913bd]
+  - ai@6.0.0-beta.110
+
+## 2.0.0-beta.109
+
+### Patch Changes
+
+- Updated dependencies [79ba99f]
+  - ai@6.0.0-beta.109
+
+## 2.0.0-beta.108
+
+### Patch Changes
+
+- ai@6.0.0-beta.108
+
+## 2.0.0-beta.107
+
+### Patch Changes
+
+- Updated dependencies [8445d70]
+  - ai@6.0.0-beta.107
+
+## 2.0.0-beta.106
+
+### Patch Changes
+
+- ai@6.0.0-beta.106
+
+## 2.0.0-beta.105
+
+### Patch Changes
+
+- ai@6.0.0-beta.105
+
+## 2.0.0-beta.104
+
+### Patch Changes
+
+- Updated dependencies [2d166e4]
+  - ai@6.0.0-beta.104
+
+## 2.0.0-beta.103
+
+### Patch Changes
+
+- ai@6.0.0-beta.103
+
 ## 2.0.0-beta.102
 
 ### Patch Changes