+++ packages/xai/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/xai
 
+## 3.0.0-beta.41
+
+### Patch Changes
+
+- 248b540: fix(xai): responses model fixes
+
+## 3.0.0-beta.40
+
+### Patch Changes
+
+- Updated dependencies [db913bd]
+  - @ai-sdk/provider@3.0.0-beta.17
+  - @ai-sdk/openai-compatible@2.0.0-beta.35
+  - @ai-sdk/provider-utils@4.0.0-beta.34
+
+## 3.0.0-beta.39
+
+### Patch Changes
+
+- 4b4c37b: fix(xai): add cache input tokens
+
 ## 3.0.0-beta.38
 
 ### Patch Changes